+++ src/main/java/com/example/Sender/repository/LetterRepository.java
@@ -4,5 +4,5 @@ package com.example.Sender.repository;
 import com.example.Sender.models.Letter;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LetterRepository extends CrudRepository<Letter, Integer> {
+public interface LetterRepository extends CrudRepository<Letter, Long> {
 }
